refactor: extract MIME type correction logic to shared utility
Create `src/utils/files/getCorrectMimeType.ts` to centralize the MIME type mapping and correction function. Remove duplicated code from `accountingAdjustment.ts` and update `applyTabClients.ts` to use the imported utility, improving maintainability and reducing code duplication.
